requirements.txt
anyio==4.0.0
api-client==0.1
astroid==3.2.4
bson==0.5.10
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dill==0.3.8
dnspython==2.6.1
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
git+ssh://git@github.com/akchau/mongo_db_client.git
packaging==24.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==1.10.17
pylint==3.2.7
pymongo==4.8.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.5
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.0.7
